// src/registry/lib/datasource-types.ts
export interface TableInfo {
  id: string
  name?: string
  title?: string
  tableMajorType?: string  // 表类型，如 'device'
  [key: string]: any
}